Move tasks to a separate module to prevent import cycles
