+++ raiden/message_handler.py
@@ -11,12 +11,12 @@ from raiden.messages import (
     RevealSecret,
     SecretRequest,
     Unlock,
+    lockedtransfersigned_from_message,
 )
 from raiden.raiden_service import RaidenService
 from raiden.routing import get_best_routes
 from raiden.transfer import views
 from raiden.transfer.architecture import StateChange
-from raiden.transfer.mediated_transfer.state import lockedtransfersigned_from_message
 from raiden.transfer.mediated_transfer.state_change import (
     ReceiveLockExpired,
     ReceiveSecretRequest,
