+++ raiden/raiden_service.py
@@ -37,6 +37,7 @@ from raiden.messages import (
     RequestMonitoring,
     SignedMessage,
     UpdatePFS,
+    lockedtransfersigned_from_message,
     message_from_sendevent,
 )
 from raiden.network.blockchain_service import BlockChainService
@@ -49,10 +50,7 @@ from raiden.tasks import AlarmTask
 from raiden.transfer import channel, node, views
 from raiden.transfer.architecture import Event as RaidenEvent, StateChange
 from raiden.transfer.mediated_transfer.events import SendLockedTransfer
-from raiden.transfer.mediated_transfer.state import (
-    TransferDescriptionWithSecretState,
-    lockedtransfersigned_from_message,
-)
+from raiden.transfer.mediated_transfer.state import TransferDescriptionWithSecretState
 from raiden.transfer.mediated_transfer.state_change import (
     ActionInitInitiator,
     ActionInitMediator,
